Move all console command classes into a more specific namespace
