+++ app/Console/Command/AbstractFireProjectedToJSONLDCommand.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace CultuurNet\UDB3\Console;
+namespace CultuurNet\UDB3\Console\Command;
 
 use Broadway\Domain\DomainEventStream;
 use Broadway\EventHandling\EventBus;
