+++ app/Console/Command/ChangeOfferOwnerInBulk.php
@@ -2,21 +2,24 @@
 
 declare(strict_types=1);
 
-namespace CultuurNet\UDB3\Console;
+namespace CultuurNet\UDB3\Console\Command;
 
 use Broadway\CommandHandling\CommandBus;
-use CultuurNet\UDB3\Organizer\Commands\ChangeOwner;
+use CultuurNet\UDB3\Offer\Commands\ChangeOwner;
 use CultuurNet\UDB3\Security\ResourceOwner\ResourceOwnerQuery;
-use CultuurNet\UDB3\StringLiteral;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Logger\ConsoleLogger;
 use Symfony\Component\Console\Output\OutputInterface;
 use Throwable;
+use CultuurNet\UDB3\StringLiteral;
 
-final class ChangeOrganizerOwnerInBulk extends AbstractCommand
+class ChangeOfferOwnerInBulk extends AbstractCommand
 {
-    private ResourceOwnerQuery $permissionQuery;
+    /**
+     * @var ResourceOwnerQuery
+     */
+    private $permissionQuery;
 
     public function __construct(
         CommandBus $commandBus,
@@ -28,8 +31,8 @@ final class ChangeOrganizerOwnerInBulk extends AbstractCommand
 
     public function configure(): void
     {
-        $this->setName('organizer:change-owner-bulk');
-        $this->setDescription('Change the owner of multiple organizers to a new user id');
+        $this->setName('offer:change-owner-bulk');
+        $this->setDescription('Change the owner of multiple offers to a new user id');
         $this->addArgument('originalOwnerId', InputArgument::REQUIRED, 'Id of the original owner');
         $this->addArgument(
             'newOwnerId',
@@ -51,24 +54,24 @@ final class ChangeOrganizerOwnerInBulk extends AbstractCommand
 
         $success = 0;
         $errors = 0;
-        foreach ($this->permissionQuery->getEditableResourceIds(new StringLiteral($originalOwnerId)) as $editableOrganizers) {
-            $organizerId = $editableOrganizers->toNative();
+        foreach ($this->permissionQuery->getEditableResourceIds(new StringLiteral($originalOwnerId)) as $editableOffer) {
+            $offerId = $editableOffer->toNative();
             try {
                 $this->commandBus->dispatch(
                     new ChangeOwner(
-                        $organizerId,
+                        $offerId,
                         $newOwnerId
                     )
                 );
                 $logger->info(
-                    'Successfully changed owner of organizer "' . $organizerId . '" to user with id "' . $newOwnerId . '"'
+                    'Successfully changed owner of offer "' . $offerId . '" to user with id "' . $newOwnerId . '"'
                 );
                 $success++;
             } catch (Throwable $t) {
                 $logger->error(
                     sprintf(
-                        'An error occurred while changing owner of organizer "%s": %s with message %s',
-                        $organizerId,
+                        'An error occurred while changing owner of offer "%s": %s with message %s',
+                        $offerId,
                         get_class($t),
                         $t->getMessage()
                     )
@@ -77,10 +80,10 @@ final class ChangeOrganizerOwnerInBulk extends AbstractCommand
             }
         }
 
-        $logger->info('Successfully changed owner of ' . $success . ' organizers to user with id "' . $newOwnerId . '"');
+        $logger->info('Successfully changed owner ' . $success . ' offers to user with id "' . $newOwnerId . '"');
 
         if ($errors) {
-            $logger->error('Failed to change owner of ' . $errors . ' organizers');
+            $logger->error('Failed to change owner of ' . $errors . ' offers');
         }
 
         return $errors > 0 ? 1 : 0;
