+++ app/Console/Command/ChangeOrganizerOwnerInBulk.php
@@ -2,24 +2,21 @@
 
 declare(strict_types=1);
 
-namespace CultuurNet\UDB3\Console;
+namespace CultuurNet\UDB3\Console\Command;
 
 use Broadway\CommandHandling\CommandBus;
-use CultuurNet\UDB3\Offer\Commands\ChangeOwner;
+use CultuurNet\UDB3\Organizer\Commands\ChangeOwner;
 use CultuurNet\UDB3\Security\ResourceOwner\ResourceOwnerQuery;
+use CultuurNet\UDB3\StringLiteral;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Logger\ConsoleLogger;
 use Symfony\Component\Console\Output\OutputInterface;
 use Throwable;
-use CultuurNet\UDB3\StringLiteral;
 
-class ChangeOfferOwnerInBulk extends AbstractCommand
+final class ChangeOrganizerOwnerInBulk extends AbstractCommand
 {
-    /**
-     * @var ResourceOwnerQuery
-     */
-    private $permissionQuery;
+    private ResourceOwnerQuery $permissionQuery;
 
     public function __construct(
         CommandBus $commandBus,
@@ -31,8 +28,8 @@ class ChangeOfferOwnerInBulk extends AbstractCommand
 
     public function configure(): void
     {
-        $this->setName('offer:change-owner-bulk');
-        $this->setDescription('Change the owner of multiple offers to a new user id');
+        $this->setName('organizer:change-owner-bulk');
+        $this->setDescription('Change the owner of multiple organizers to a new user id');
         $this->addArgument('originalOwnerId', InputArgument::REQUIRED, 'Id of the original owner');
         $this->addArgument(
             'newOwnerId',
@@ -54,24 +51,24 @@ class ChangeOfferOwnerInBulk extends AbstractCommand
 
         $success = 0;
         $errors = 0;
-        foreach ($this->permissionQuery->getEditableResourceIds(new StringLiteral($originalOwnerId)) as $editableOffer) {
-            $offerId = $editableOffer->toNative();
+        foreach ($this->permissionQuery->getEditableResourceIds(new StringLiteral($originalOwnerId)) as $editableOrganizers) {
+            $organizerId = $editableOrganizers->toNative();
             try {
                 $this->commandBus->dispatch(
                     new ChangeOwner(
-                        $offerId,
+                        $organizerId,
                         $newOwnerId
                     )
                 );
                 $logger->info(
-                    'Successfully changed owner of offer "' . $offerId . '" to user with id "' . $newOwnerId . '"'
+                    'Successfully changed owner of organizer "' . $organizerId . '" to user with id "' . $newOwnerId . '"'
                 );
                 $success++;
             } catch (Throwable $t) {
                 $logger->error(
                     sprintf(
-                        'An error occurred while changing owner of offer "%s": %s with message %s',
-                        $offerId,
+                        'An error occurred while changing owner of organizer "%s": %s with message %s',
+                        $organizerId,
                         get_class($t),
                         $t->getMessage()
                     )
@@ -80,10 +77,10 @@ class ChangeOfferOwnerInBulk extends AbstractCommand
             }
         }
 
-        $logger->info('Successfully changed owner ' . $success . ' offers to user with id "' . $newOwnerId . '"');
+        $logger->info('Successfully changed owner of ' . $success . ' organizers to user with id "' . $newOwnerId . '"');
 
         if ($errors) {
-            $logger->error('Failed to change owner of ' . $errors . ' offers');
+            $logger->error('Failed to change owner of ' . $errors . ' organizers');
         }
 
         return $errors > 0 ? 1 : 0;
