+++ app/Console/Command/ConsumeCommand.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace CultuurNet\UDB3\Console;
+namespace CultuurNet\UDB3\Console\Command;
 
 use Closure;
 use CultuurNet\UDB3\Broadway\AMQP\ConsumerInterface;
