+++ app/Console/Command/GeocodeOrganizerCommand.php
@@ -2,21 +2,21 @@
 
 declare(strict_types=1);
 
-namespace CultuurNet\UDB3\Console;
+namespace CultuurNet\UDB3\Console\Command;
 
 use CultuurNet\UDB3\Address\Address;
 use CultuurNet\UDB3\Json;
-use CultuurNet\UDB3\Place\Commands\UpdateGeoCoordinatesFromAddress;
+use CultuurNet\UDB3\Organizer\Commands\UpdateGeoCoordinatesFromAddress;
 use Symfony\Component\Console\Output\OutputInterface;
 
-class GeocodePlaceCommand extends AbstractGeocodeCommand
+class GeocodeOrganizerCommand extends AbstractGeocodeCommand
 {
     public function configure(): void
     {
         parent::configure();
         $this
-            ->setName('place:geocode')
-            ->setDescription('Geocode places with missing or outdated coordinates.');
+            ->setName('organizer:geocode')
+            ->setDescription('Geocode organizers with missing or outdated coordinates.');
     }
 
     protected function getQueryForMissingCoordinates(): string
@@ -24,11 +24,12 @@ class GeocodePlaceCommand extends AbstractGeocodeCommand
         return '_exists_:address NOT(_exists_:geo OR workflowStatus:DELETED OR workflowStatus:REJECTED)';
     }
 
-    protected function dispatchGeocodingCommand(string $placeId, OutputInterface $output): void
+    protected function dispatchGeocodingCommand(string $organizerId, OutputInterface $output): void
     {
-        $document = $this->getDocument($placeId);
+        $document = $this->getDocument($organizerId);
+
         if (is_null($document)) {
-            $output->writeln("Skipping {$placeId}. (Could not find JSON-LD in local repository.)");
+            $output->writeln("Skipping {$organizerId}. (Could not find JSON-LD in local repository.)");
             return;
         }
 
@@ -37,10 +38,10 @@ class GeocodePlaceCommand extends AbstractGeocodeCommand
         $addressLanguage = $jsonLd->mainLanguage ?? 'nl';
 
         if (!isset($jsonLd['address'][$addressLanguage])) {
-            // Some places have an address in another language then the main language or `nl`
+            // Some organizers have an address in another language then the main language or `nl`
             $addressLanguage = array_key_first($jsonLd['address']);
             if ($addressLanguage === null) {
-                $output->writeln("Skipping {$placeId}. (JSON-LD does not contain an address for {$addressLanguage}.)");
+                $output->writeln("Skipping {$organizerId}. (JSON-LD does not contain an address for {$addressLanguage}.)");
                 return;
             }
         }
@@ -48,14 +49,14 @@ class GeocodePlaceCommand extends AbstractGeocodeCommand
         try {
             $address = Address::deserialize($jsonLd['address'][$addressLanguage]);
         } catch (\Exception $e) {
-            $output->writeln("Skipping {$placeId}. (JSON-LD address for {$addressLanguage} could not be parsed.)");
+            $output->writeln("Skipping {$organizerId}. (JSON-LD address for {$addressLanguage} could not be parsed.)");
             return;
         }
 
-        $output->writeln("Dispatching geocode command for place {$placeId}.");
+        $output->writeln("Dispatching geocode command for organizer {$organizerId}.");
 
         $this->commandBus->dispatch(
-            new UpdateGeoCoordinatesFromAddress($placeId, $address)
+            new UpdateGeoCoordinatesFromAddress($organizerId, $address)
         );
     }
 }
