+++ app/Console/Command/RemoveLabelOffer.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace CultuurNet\UDB3\Console;
+namespace CultuurNet\UDB3\Console\Command;
 
 use CultuurNet\UDB3\Offer\Commands\RemoveLabel;
 use Symfony\Component\Console\Input\InputArgument;
