+++ app/Console/Command/RemoveLabelOrganizer.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace CultuurNet\UDB3\Console;
+namespace CultuurNet\UDB3\Console\Command;
 
 use CultuurNet\UDB3\Organizer\Commands\RemoveLabel;
 use Symfony\Component\Console\Input\InputArgument;
