+++ bin/udb3.php
@@ -7,32 +7,32 @@ use CultuurNet\UDB3\Event\ReadModel\Relations\EventRelationsRepository;
 use CultuurNet\UDB3\Offer\OfferType;
 use CultuurNet\UDB3\Organizer\WebsiteNormalizer;
 use CultuurNet\UDB3\ApiName;
-use CultuurNet\UDB3\Console\ChangeOfferOwner;
-use CultuurNet\UDB3\Console\ChangeOfferOwnerInBulk;
-use CultuurNet\UDB3\Console\ChangeOrganizerOwner;
-use CultuurNet\UDB3\Console\ChangeOrganizerOwnerInBulk;
-use CultuurNet\UDB3\Console\ConsumeCommand;
-use CultuurNet\UDB3\Console\EventAncestorsCommand;
-use CultuurNet\UDB3\Console\FireProjectedToJSONLDCommand;
-use CultuurNet\UDB3\Console\FireProjectedToJSONLDForRelationsCommand;
-use CultuurNet\UDB3\Console\GeocodeEventCommand;
-use CultuurNet\UDB3\Console\GeocodeOrganizerCommand;
-use CultuurNet\UDB3\Console\GeocodePlaceCommand;
-use CultuurNet\UDB3\Console\ImportOfferAutoClassificationLabels;
-use CultuurNet\UDB3\Console\ProcessDuplicatePlaces;
-use CultuurNet\UDB3\Console\PurgeModelCommand;
-use CultuurNet\UDB3\Console\ReindexEventsWithRecommendations;
-use CultuurNet\UDB3\Console\ReindexOffersWithPopularityScore;
-use CultuurNet\UDB3\Console\RemoveFacilitiesFromPlace;
-use CultuurNet\UDB3\Console\RemoveLabelOffer;
-use CultuurNet\UDB3\Console\RemoveLabelOrganizer;
-use CultuurNet\UDB3\Console\ReplaceNewsArticlePublisher;
-use CultuurNet\UDB3\Console\ReplayCommand;
-use CultuurNet\UDB3\Console\UpdateBookingAvailabilityCommand;
-use CultuurNet\UDB3\Console\UpdateEventsAttendanceMode;
-use CultuurNet\UDB3\Console\UpdateOfferStatusCommand;
-use CultuurNet\UDB3\Console\UpdateUniqueLabels;
-use CultuurNet\UDB3\Console\UpdateUniqueOrganizers;
+use CultuurNet\UDB3\Console\Command\ChangeOfferOwner;
+use CultuurNet\UDB3\Console\Command\ChangeOfferOwnerInBulk;
+use CultuurNet\UDB3\Console\Command\ChangeOrganizerOwner;
+use CultuurNet\UDB3\Console\Command\ChangeOrganizerOwnerInBulk;
+use CultuurNet\UDB3\Console\Command\ConsumeCommand;
+use CultuurNet\UDB3\Console\Command\EventAncestorsCommand;
+use CultuurNet\UDB3\Console\Command\FireProjectedToJSONLDCommand;
+use CultuurNet\UDB3\Console\Command\FireProjectedToJSONLDForRelationsCommand;
+use CultuurNet\UDB3\Console\Command\GeocodeEventCommand;
+use CultuurNet\UDB3\Console\Command\GeocodeOrganizerCommand;
+use CultuurNet\UDB3\Console\Command\GeocodePlaceCommand;
+use CultuurNet\UDB3\Console\Command\ImportOfferAutoClassificationLabels;
+use CultuurNet\UDB3\Console\Command\ProcessDuplicatePlaces;
+use CultuurNet\UDB3\Console\Command\PurgeModelCommand;
+use CultuurNet\UDB3\Console\Command\ReindexEventsWithRecommendations;
+use CultuurNet\UDB3\Console\Command\ReindexOffersWithPopularityScore;
+use CultuurNet\UDB3\Console\Command\RemoveFacilitiesFromPlace;
+use CultuurNet\UDB3\Console\Command\RemoveLabelOffer;
+use CultuurNet\UDB3\Console\Command\RemoveLabelOrganizer;
+use CultuurNet\UDB3\Console\Command\ReplaceNewsArticlePublisher;
+use CultuurNet\UDB3\Console\Command\ReplayCommand;
+use CultuurNet\UDB3\Console\Command\UpdateBookingAvailabilityCommand;
+use CultuurNet\UDB3\Console\Command\UpdateEventsAttendanceMode;
+use CultuurNet\UDB3\Console\Command\UpdateOfferStatusCommand;
+use CultuurNet\UDB3\Console\Command\UpdateUniqueLabels;
+use CultuurNet\UDB3\Console\Command\UpdateUniqueOrganizers;
 use CultuurNet\UDB3\Silex\Container\HybridContainerApplication;
 use CultuurNet\UDB3\Error\CliErrorHandlerProvider;
 use CultuurNet\UDB3\Error\ErrorLogger;
